+++ x/evm/precompile/funtoken.go
@@ -11,6 +11,7 @@ import (
 	"github.com/ethereum/go-ethereum/core/vm"
 
 	"github.com/NibiruChain/nibiru/v2/app/keepers"
+	"github.com/NibiruChain/nibiru/v2/eth"
 	"github.com/NibiruChain/nibiru/v2/x/evm"
 	"github.com/NibiruChain/nibiru/v2/x/evm/embeds"
 	evmkeeper "github.com/NibiruChain/nibiru/v2/x/evm/keeper"
@@ -22,18 +23,6 @@ var _ vm.PrecompiledContract = (*precompileFunToken)(nil)
 // enables transfers of ERC20 tokens to "nibi" addresses as bank coins
 // using the ERC20's `FunToken` mapping.
 var PrecompileAddr_FunToken = gethcommon.HexToAddress("0x0000000000000000000000000000000000000800")
-
-const (
-	// FunTokenGasLimitBankSend consists of gas for 3 calls:
-	// 1. transfer erc20 from sender to module
-	//    ~60_000 gas for regular erc20 transfer (our own ERC20Minter contract)
-	//    could be higher for user created contracts, let's cap with 200_000
-	// 2. mint native coin (made from erc20) or burn erc20 token (made from coin)
-	//	  ~60_000 gas for either mint or burn
-	// 3. send from module to account:
-	//	  ~65_000 gas (bank send)
-	FunTokenGasLimitBankSend uint64 = 400_000
-)
 
 func (p precompileFunToken) Address() gethcommon.Address {
 	return PrecompileAddr_FunToken
@@ -49,7 +38,10 @@ func (p precompileFunToken) ABI() *gethabi.ABI {
 }
 
 const (
-	FunTokenMethod_BankSend PrecompileMethod = "bankSend"
+	FunTokenMethod_sendToBank  PrecompileMethod = "sendToBank"
+	FunTokenMethod_balance     PrecompileMethod = "balance"
+	FunTokenMethod_bankBalance PrecompileMethod = "bankBalance"
+	FunTokenMethod_whoAmI      PrecompileMethod = "whoAmI"
 )
 
 // Run runs the precompiled contract
@@ -70,11 +62,17 @@ func (p precompileFunToken) Run(
 
 	method := startResult.Method
 	switch PrecompileMethod(method.Name) {
-	case FunTokenMethod_BankSend:
-		bz, err = p.bankSend(startResult, contract.CallerAddress, readonly)
+	case FunTokenMethod_sendToBank:
+		bz, err = p.sendToBank(startResult, contract.CallerAddress, readonly)
+	case FunTokenMethod_balance:
+		bz, err = p.balance(startResult, contract)
+	case FunTokenMethod_bankBalance:
+		bz, err = p.bankBalance(startResult, contract)
+	case FunTokenMethod_whoAmI:
+		bz, err = p.whoAmI(startResult, contract)
 	default:
 		// Note that this code path should be impossible to reach since
-		// "DecomposeInput" parses methods directly from the ABI.
+		// "[decomposeInput]" parses methods directly from the ABI.
 		err = fmt.Errorf("invalid method called with name \"%s\"", method.Name)
 		return
 	}
@@ -97,18 +95,18 @@ type precompileFunToken struct {
 	evmKeeper *evmkeeper.Keeper
 }
 
-// bankSend: Implements "IFunToken.bankSend"
+// sendToBank: Implements "IFunToken.sendToBank"
 //
 // The "args" populate the following function signature in Solidity:
 //
 //	```solidity
-//	/// @dev bankSend sends ERC20 tokens as coins to a Nibiru base account
+//	/// @dev sendToBank sends ERC20 tokens as coins to a Nibiru base account
 //	/// @param erc20 the address of the ERC20 token contract
 //	/// @param amount the amount of tokens to send
 //	/// @param to the receiving Nibiru base account address as a string
-//	function bankSend(address erc20, uint256 amount, string memory to) external;
+//	function sendToBank(address erc20, uint256 amount, string memory to) external;
 //	```
-func (p precompileFunToken) bankSend(
+func (p precompileFunToken) sendToBank(
 	startResult OnRunStartResult,
 	caller gethcommon.Address,
 	readOnly bool,
@@ -118,7 +116,7 @@ func (p precompileFunToken) bankSend(
 		return nil, err
 	}
 
-	erc20, amount, to, err := p.decomposeBankSendArgs(args)
+	erc20, amount, to, err := p.parseArgsSendToBank(args)
 	if err != nil {
 		return
 	}
@@ -206,38 +204,299 @@ func (p precompileFunToken) bankSend(
 	}
 
 	// TODO: UD-DEBUG: feat: Emit EVM events
+	// TODO: emit event for balance change of sender
+	// TODO: emit event for balance change of recipient
 
 	return method.Outputs.Pack(gotAmount)
 }
 
-func (p precompileFunToken) decomposeBankSendArgs(args []any) (
+func (p precompileFunToken) parseArgsSendToBank(args []any) (
 	erc20 gethcommon.Address,
 	amount *big.Int,
 	to string,
 	err error,
 ) {
-	if e := assertNumArgs(len(args), 3); e != nil {
+	if e := assertNumArgs(args, 3); e != nil {
 		err = e
 		return
 	}
 
-	erc20, ok := args[0].(gethcommon.Address)
+	argIdx := 0
+	erc20, ok := args[argIdx].(gethcommon.Address)
 	if !ok {
-		err = fmt.Errorf("type validation for failed for (address erc20) argument")
+		err = ErrArgTypeValidation("address erc20", args[argIdx])
 		return
 	}
 
-	amount, ok = args[1].(*big.Int)
+	argIdx++
+	amount, ok = args[argIdx].(*big.Int)
 	if !ok {
-		err = fmt.Errorf("type validation for failed for (uint256 amount) argument")
+		err = ErrArgTypeValidation("uint256 amount", args[argIdx])
 		return
 	}
 
-	to, ok = args[2].(string)
+	argIdx++
+	to, ok = args[argIdx].(string)
 	if !ok {
-		err = fmt.Errorf("type validation for failed for (string to) argument")
+		err = ErrArgTypeValidation("string to", args[argIdx])
 		return
 	}
 
 	return
+}
+
+// balance: Implements "IFunToken.balance"
+//
+// The "args" populate the following function signature in Solidity:
+//
+//	```solidity
+//	function balance(
+//	    address who,
+//	    address funtoken
+//	)
+//	    external
+//	    returns (
+//	        uint256 erc20Balance,
+//	        uint256 bankBalance,
+//	        FunToken memory token,
+//	        NibiruAccount memory whoAddrs
+//	    );
+//	```
+func (p precompileFunToken) balance(
+	start OnRunStartResult,
+	contract *vm.Contract,
+) (bz []byte, err error) {
+	method, args, ctx := start.Method, start.Args, start.CacheCtx
+	defer func() {
+		if err != nil {
+			err = ErrMethodCalled(method, err)
+		}
+	}()
+	if err := assertContractQuery(contract); err != nil {
+		return bz, err
+	}
+
+	addrEth, addrBech32, funtoken, err := p.parseArgsBalance(args, ctx)
+	if err != nil {
+		err = ErrInvalidArgs(err)
+		return
+	}
+
+	erc20Bal, err := p.evmKeeper.ERC20().BalanceOf(funtoken.Erc20Addr.Address, addrEth, ctx)
+	if err != nil {
+		return
+	}
+	bankBal := p.evmKeeper.Bank.GetBalance(ctx, addrBech32, funtoken.BankDenom).Amount.BigInt()
+
+	return method.Outputs.Pack([]any{
+		erc20Bal,
+		bankBal,
+		struct {
+			Erc20     gethcommon.Address `json:"erc20"`
+			BankDenom string             `json:"bankDenom"`
+		}{
+			Erc20:     funtoken.Erc20Addr.Address,
+			BankDenom: funtoken.BankDenom,
+		},
+		struct {
+			EthAddr    gethcommon.Address `json:"ethAddr"`
+			Bech32Addr string             `json:"bech32Addr"`
+		}{
+			EthAddr:    addrEth,
+			Bech32Addr: addrBech32.String(),
+		},
+	}...)
+}
+
+func (p precompileFunToken) parseArgsBalance(args []any, ctx sdk.Context) (
+	addrEth gethcommon.Address,
+	addrBech32 sdk.AccAddress,
+	funtoken evm.FunToken,
+	err error,
+) {
+	if e := assertNumArgs(args, 2); e != nil {
+		err = e
+		return
+	}
+
+	argIdx := 0
+	who, ok := args[argIdx].(gethcommon.Address)
+	if !ok {
+		err = ErrArgTypeValidation("bytes who", args[argIdx])
+		return
+	}
+	req := &evm.QueryEthAccountRequest{Address: who.Hex()}
+	_, e := req.Validate()
+	if e != nil {
+		err = e
+		return
+	}
+	addrEth = gethcommon.HexToAddress(req.Address)
+	addrBech32 = eth.EthAddrToNibiruAddr(addrEth)
+
+	argIdx++
+	funtokenErc20, ok := args[argIdx].(gethcommon.Address)
+	if !ok {
+		err = ErrArgTypeValidation("bytes funtoken", args[argIdx])
+		return
+	}
+	resp, e := p.evmKeeper.FunTokenMapping(ctx, &evm.QueryFunTokenMappingRequest{
+		Token: funtokenErc20.Hex(),
+	})
+	if e != nil {
+		err = e
+		return
+	}
+
+	return addrEth, addrBech32, *resp.FunToken, nil
+}
+
+// bankBalance: Implements "IFunToken.bankBalance"
+//
+// The "args" populate the following function signature in Solidity:
+//
+//	```solidity
+//	function bankBalance(
+//	    address who,
+//	    string calldata bankDenom
+//	) external returns (uint256 bankBalance, NibiruAccount memory whoAddrs);
+//	```
+func (p precompileFunToken) bankBalance(
+	start OnRunStartResult,
+	contract *vm.Contract,
+) (bz []byte, err error) {
+	method, args, ctx := start.Method, start.Args, start.CacheCtx
+	defer func() {
+		if err != nil {
+			err = ErrMethodCalled(method, err)
+		}
+	}()
+	if err := assertContractQuery(contract); err != nil {
+		return bz, err
+	}
+
+	addrEth, addrBech32, bankDenom, err := p.parseArgsBankBalance(args)
+	bankBal := p.evmKeeper.Bank.GetBalance(ctx, addrBech32, bankDenom).Amount.BigInt()
+
+	return method.Outputs.Pack([]any{
+		bankBal,
+		struct {
+			EthAddr    gethcommon.Address `json:"ethAddr"`
+			Bech32Addr string             `json:"bech32Addr"`
+		}{
+			EthAddr:    addrEth,
+			Bech32Addr: addrBech32.String(),
+		},
+	}...)
+}
+
+func (p precompileFunToken) parseArgsBankBalance(args []any) (
+	addrEth gethcommon.Address,
+	addrBech32 sdk.AccAddress,
+	bankDenom string,
+	err error,
+) {
+	if e := assertNumArgs(args, 2); e != nil {
+		err = e
+		return
+	}
+
+	argIdx := 0
+	who, ok := args[argIdx].(gethcommon.Address)
+	if !ok {
+		err = ErrArgTypeValidation("bytes who", args[argIdx])
+		return
+	}
+	req := &evm.QueryEthAccountRequest{Address: who.Hex()}
+	_, e := req.Validate()
+	if e != nil {
+		err = e
+		return
+	}
+	addrEth = gethcommon.HexToAddress(req.Address)
+	addrBech32 = eth.EthAddrToNibiruAddr(addrEth)
+
+	argIdx++
+	bankDenom, ok = args[argIdx].(string)
+	if !ok {
+		err = ErrArgTypeValidation("string bankDenom", args[argIdx])
+		return
+	}
+	if e := sdk.ValidateDenom(bankDenom); e != nil {
+		err = e
+		return
+	}
+
+	return addrEth, addrBech32, bankDenom, nil
+}
+
+// whoAmI: Implements "IFunToken.whoAmI"
+//
+// The "args" populate the following function signature in Solidity:
+//
+//	```solidity
+//	function whoAmI(
+//	    string calldata who
+//	) external returns (NibiruAccount memory whoAddrs);
+//	```
+func (p precompileFunToken) whoAmI(
+	start OnRunStartResult,
+	contract *vm.Contract,
+) (bz []byte, err error) {
+	method, args := start.Method, start.Args
+	defer func() {
+		if err != nil {
+			err = ErrMethodCalled(method, err)
+		}
+	}()
+	if err := assertContractQuery(contract); err != nil {
+		return bz, err
+	}
+
+	addrEth, addrBech32, err := p.parseArgsWhoAmI(args)
+	if err != nil {
+		return bz, err
+	}
+
+	return method.Outputs.Pack([]any{
+		struct {
+			EthAddr    gethcommon.Address `json:"ethAddr"`
+			Bech32Addr string             `json:"bech32Addr"`
+		}{
+			EthAddr:    addrEth,
+			Bech32Addr: addrBech32.String(),
+		},
+	}...)
+}
+
+func (p precompileFunToken) parseArgsWhoAmI(args []any) (
+	addrEth gethcommon.Address,
+	addrBech32 sdk.AccAddress,
+	err error,
+) {
+	if e := assertNumArgs(args, 1); e != nil {
+		err = e
+		return
+	}
+
+	argIdx := 0
+	who, ok := args[argIdx].(string)
+	if !ok {
+		err = ErrArgTypeValidation("string calldata who", args[argIdx])
+		return
+	}
+	req := &evm.QueryEthAccountRequest{Address: who}
+	isBech32, e := req.Validate()
+	if e != nil {
+		err = e
+		return
+	}
+	if isBech32 {
+		addrBech32 = sdk.MustAccAddressFromBech32(req.Address)
+		addrEth = eth.NibiruAddrToEthAddr(addrBech32)
+	} else {
+		addrEth = gethcommon.HexToAddress(req.Address)
+		addrBech32 = eth.EthAddrToNibiruAddr(addrEth)
+	}
+	return addrEth, addrBech32, nil
 }