+++ x/evm/precompile/wasm_parse.go
@@ -51,7 +51,7 @@ func parseFundsArg(arg any) (funds sdk.Coins, err error) {
 }
 
 // Parses [sdk.AccAddress] from a "string" solidity argument:
-func parseContractAddrArg(arg any) (addr sdk.AccAddress, err error) {
+func parseArgContractAddr(arg any) (addr sdk.AccAddress, err error) {
 	addrStr, ok := arg.(string)
 	if !ok {
 		err = ErrArgTypeValidation("string contractAddr", arg)
@@ -67,11 +67,11 @@ func parseContractAddrArg(arg any) (addr sdk.AccAddress, err error) {
 	return addr, nil
 }
 
-func (p precompileWasm) parseInstantiateArgs(args []any, sender string) (
+func (p precompileWasm) parseArgsWasmInstantiate(args []any, sender string) (
 	txMsg wasm.MsgInstantiateContract,
 	err error,
 ) {
-	if e := assertNumArgs(len(args), 5); e != nil {
+	if e := assertNumArgs(args, 5); e != nil {
 		err = e
 		return
 	}
@@ -124,13 +124,13 @@ func (p precompileWasm) parseInstantiateArgs(args []any, sender string) (
 	return txMsg, txMsg.ValidateBasic()
 }
 
-func (p precompileWasm) parseExecuteArgs(args []any) (
+func (p precompileWasm) parseArgsWasmExecute(args []any) (
 	wasmContract sdk.AccAddress,
 	msgArgs []byte,
 	funds sdk.Coins,
 	err error,
 ) {
-	if e := assertNumArgs(len(args), 3); e != nil {
+	if e := assertNumArgs(args, 3); e != nil {
 		err = e
 		return
 	}
@@ -174,18 +174,18 @@ func (p precompileWasm) parseExecuteArgs(args []any) (
 	return contractAddr, msgArgs, funds, nil
 }
 
-func (p precompileWasm) parseQueryArgs(args []any) (
+func (p precompileWasm) parseArgsWasmQuery(args []any) (
 	wasmContract sdk.AccAddress,
 	req wasm.RawContractMessage,
 	err error,
 ) {
-	if e := assertNumArgs(len(args), 2); e != nil {
+	if e := assertNumArgs(args, 2); e != nil {
 		err = e
 		return
 	}
 
 	argsIdx := 0
-	wasmContract, e := parseContractAddrArg(args[argsIdx])
+	wasmContract, e := parseArgContractAddr(args[argsIdx])
 	if e != nil {
 		err = e
 		return
@@ -206,7 +206,7 @@ func (p precompileWasm) parseQueryArgs(args []any) (
 	return wasmContract, req, nil
 }
 
-func (p precompileWasm) parseExecuteMultiArgs(args []any) (
+func (p precompileWasm) parseArgsWasmExecuteMulti(args []any) (
 	wasmExecMsgs []struct {
 		ContractAddr string `json:"contractAddr"`
 		MsgArgs      []byte `json:"msgArgs"`
@@ -217,7 +217,7 @@ func (p precompileWasm) parseExecuteMultiArgs(args []any) (
 	},
 	err error,
 ) {
-	if e := assertNumArgs(len(args), 1); e != nil {
+	if e := assertNumArgs(args, 1); e != nil {
 		err = e
 		return
 	}